+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ### June
 
-* June 8 - Add country search for Developers - #468 @troyizzle
+* June 8 - Add text search for developers (for paid accounts) - #470 @troyizzle
+* June 8 - Add country search for developers - #468 @troyizzle
 * June 8 - Refactor non-reusable View Components under model named module - #456, #462, #471, #476, #478, #480 @metamoni
 
 ### May